A simple working Twitter Bot
This is an updated version, with simplification of the twitterer class, and the addition of the webscrape class. Webscrape is currently not used by the Twitter bot but it will be implemented later.  The twitter bot currently runs for 48 hours and will tweet 24 times, randomly out of a list of 6 hard coded tweets, once every 2 hours.